+++ src/operators/query/bitwise/index.ts
@@ -1,0 +1,4 @@
+export { $bitsAllClear } from "./bitsAllClear";
+export { $bitsAllSet } from "./bitsAllSet";
+export { $bitsAnyClear } from "./bitsAnyClear";
+export { $bitsAnySet } from "./bitsAnySet";