+++ src/operators/query/index.ts
@@ -1,4 +1,5 @@
 export * from "./array";
+export * from "./bitwise";
 export * from "./comparison";
 export * from "./element";
 export * from "./evaluation";